Create Bakery (Dictionaries)
You will receive a single line containing some food (keys) and quantities (values). They will be separated by a single space (the first element is the key, the second – is the value, and so on). Create a dictionary with all the keys and values and print it on the console.